refactor(new-request-notification): create notify route controller
