+++ src/controllers/notificationController.js
@@ -1,3 +1,4 @@
+import webPush from 'web-push';
 import Service from '../services/notificationService';
 import Response from '../utils/response';
 
@@ -21,5 +22,25 @@ export default class Notification {
     }
     response.setSuccess(200, 'You have successfully unsubscribed from email notifications');
     return response.send(res);
+  }
+
+  /**
+ *
+ * @param {obj} body => message to be sent to front end and the suvscription received from front end
+ * @returns {obj} returns an object
+ */
+  static async notify(body) {
+    const { PUBLIC_VAPID_KEY, PRIVATE_VAPID_KEY } = process.env;
+    await webPush.setVapidDetails('mailto:walesadeks@gmail.com', PUBLIC_VAPID_KEY, PRIVATE_VAPID_KEY);
+    const { subscription, message } = body;
+    try {
+      const payload = JSON.stringify({
+        title: 'New Travel Request',
+        body: message
+      });
+      await webPush.sendNotification(subscription, payload);
+    } catch (error) {
+      throw new Error(error.message);
+    }
   }
 }
