+++ src/routes/api/notifications.js
@@ -1,4 +1,3 @@
-import webPush from 'web-push';
 import { Router } from 'express';
 import notificationController from '../../controllers/notificationController';
 import { authorization } from '../../middlewares/auth/auth';
@@ -7,11 +6,6 @@ import { permit } from '../../middlewares/users';
 import { roleIds } from '../../helpers/default';
 
 const router = Router();
-
-const { PUBLIC_VAPID_KEY, PRIVATE_VAPID_KEY } = process.env;
-
-webPush.setVapidDetails('mailto:walesadeks@gmail.com', PUBLIC_VAPID_KEY, PRIVATE_VAPID_KEY);
-
 
 /**
  * @swagger
@@ -33,18 +27,10 @@ webPush.setVapidDetails('mailto:walesadeks@gmail.com', PUBLIC_VAPID_KEY, PRIVATE
  *              $ref: '#/components/schemas/response'
  */
 router.post('/notifications/notify', async (req, res) => {
-  const { subscription, message } = req.body;
-  res.status(201).json({
-    success: true
-  });
   try {
-    const payload = JSON.stringify({
-      title: 'New Travel Request',
-      body: message
-    });
-    await webPush.sendNotification(subscription, payload);
+    await notificationController.notify(req.body);
   } catch (error) {
-    res.status().json({
+    res.status(500).json({
       success: false,
       message: 'Could not send push notification'
     });
